move scripts other than main to lib
