+++ lib/geoTable.js
