+++ lib/normalTable.js
